Update double.c to eliminate data segment allocations (#1)
The double pointers for each function that either point to the console FAC or the ARG space, were quite redundant, and end up allocating a word each for 12 pointers. 3 per function. 

This isn't necessary, by using macros that evaluate to the FAC as a volatile double*, the compiler will slightly more efficient code, and not require this allocation of 24 bytes of memory. 

Additional 6 more bytes of memory are delcared and not ever used. So this removes those.

And the declarations in math_private.h are not used, so there is no need to include it.